[리뷰] 장바구니 화면에서도 참조되고 있는 컴포넌트임으로 ui/component 패키지로 이동하면 어떨까요?
