+++ app/src/main/java/nextstep/shoppingcart/ui/screen/product/list/ProductListScreen.kt
@@ -36,7 +36,7 @@ import nextstep.shoppingcart.R
 import nextstep.shoppingcart.model.Product
 import nextstep.shoppingcart.ui.component.PriceText
 import nextstep.shoppingcart.ui.component.ProductImage
-import nextstep.shoppingcart.ui.screen.product.detail.ProductTitle
+import nextstep.shoppingcart.ui.component.ProductTitle
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
